Update jar to fix count, convert from int to long
git-svn-id: file:///home/svn/shoola-omero/trunk@5656 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/agents/treeviewer/ContainerCounterLoader.java
@@ -126,12 +126,12 @@ public class ContainerCounterLoader
         if (map == null) return;
         Iterator i = map.keySet().iterator();
         Long containerID;
-        Integer value;
+        Long value;
         while (i.hasNext()) {
             containerID = (Long) i.next();
-            value = (Integer) map.get(containerID);
+            value = (Long) map.get(containerID);
             viewer.setContainerCountValue(containerID.longValue(),
-                    						value.intValue());
+                    						value.longValue());
         }
     }
